+++ packages/adaptive-ui-designer-core/CHANGELOG.md
@@ -1,8 +1,49 @@
 # Change Log - @adaptive-web/adaptive-ui-designer-core
 
-This log was last generated on Mon, 24 Jun 2024 21:29:12 GMT and should not be manually modified.
+This log was last generated on Thu, 25 Jul 2024 17:29:23 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.1
+
+Thu, 25 Jul 2024 17:29:23 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.9.1
+
+## 0.3.0
+
+Fri, 19 Jul 2024 17:39:45 GMT
+
+### Minor changes
+
+- Designer: Added ability to mark a node `inline` for styling (47367562+bheston@users.noreply.github.com)
+- Bump @adaptive-web/adaptive-ui to v0.9.0
+
+## 0.2.3
+
+Mon, 15 Jul 2024 18:51:14 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.8.0
+
+## 0.2.2
+
+Thu, 11 Jul 2024 17:06:10 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.7.0
+
+## 0.2.1
+
+Tue, 09 Jul 2024 16:29:52 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.6.0
 
 ## 0.2.0