+++ packages/adaptive-ui-designer-figma/CHANGELOG.md
@@ -1,8 +1,61 @@
 # Change Log - @adaptive-web/adaptive-ui-designer-figma
 
-This log was last generated on Mon, 24 Jun 2024 21:29:12 GMT and should not be manually modified.
+This log was last generated on Thu, 25 Jul 2024 17:29:23 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.1
+
+Thu, 25 Jul 2024 17:29:23 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.9.1
+- Bump @adaptive-web/adaptive-ui-designer-core to v0.3.1
+
+## 0.4.0
+
+Fri, 19 Jul 2024 17:39:45 GMT
+
+### Minor changes
+
+- AUI: Update CLI from `tokens` to align with existing `properties` (47367562+bheston@users.noreply.github.com)
+- Designer: Added ability to mark a node `inline` for styling (47367562+bheston@users.noreply.github.com)
+- Bump @adaptive-web/adaptive-ui to v0.9.0
+- Bump @adaptive-web/adaptive-ui-designer-core to v0.3.0
+
+## 0.3.1
+
+Mon, 15 Jul 2024 18:51:14 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.8.0
+- Bump @adaptive-web/adaptive-ui-designer-core to v0.2.3
+
+## 0.3.0
+
+Thu, 11 Jul 2024 17:06:10 GMT
+
+### Minor changes
+
+- AUI: Fix global style generation (47367562+bheston@users.noreply.github.com)
+- Bump @adaptive-web/adaptive-ui to v0.7.0
+- Bump @adaptive-web/adaptive-ui-designer-core to v0.2.2
+
+## 0.2.0
+
+Tue, 09 Jul 2024 16:29:52 GMT
+
+### Minor changes
+
+- Improved Figma CLI (47367562+bheston@users.noreply.github.com)
+- Bump @adaptive-web/adaptive-ui to v0.6.0
+- Bump @adaptive-web/adaptive-ui-designer-core to v0.2.1
+
+### Patches
+
+- Adaptive UI: Anatomy CLI fixes (47367562+bheston@users.noreply.github.com)
 
 ## 0.1.2