+++ packages/adaptive-web-components/CHANGELOG.md
@@ -1,8 +1,49 @@
 # Change Log - @adaptive-web/adaptive-web-components
 
-This log was last generated on Mon, 24 Jun 2024 21:29:12 GMT and should not be manually modified.
+This log was last generated on Thu, 25 Jul 2024 17:29:23 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.4
+
+Thu, 25 Jul 2024 17:29:23 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.9.1
+
+## 0.7.3
+
+Fri, 19 Jul 2024 17:39:45 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.9.0
+
+## 0.7.2
+
+Mon, 15 Jul 2024 18:51:14 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.8.0
+
+## 0.7.1
+
+Thu, 11 Jul 2024 17:06:10 GMT
+
+### Patches
+
+- Bump @adaptive-web/adaptive-ui to v0.7.0
+
+## 0.7.0
+
+Tue, 09 Jul 2024 16:29:52 GMT
+
+### Minor changes
+
+- Adaptive UI: Migrated global disabled and focus style support from AWC to AUI (47367562+bheston@users.noreply.github.com)
+- Bump @adaptive-web/adaptive-ui to v0.6.0
 
 ## 0.6.2